Fix InstrumentDefMsg binary format to match Databento documentation
Major changes based on official Databento docs:

1. Binary field order: raw_symbol, security_update_action, and
   instrument_class come immediately after ts_recv, NOT after all numerics
2. Type corrections: channel_id, leg_count, leg_index are uint16_t not uint8_t
3. Leg field order: leg_instrument_class and leg_side before leg_price/leg_delta,
   then ratio fields, then leg_underlying_id at end
4. String field order after numeric fields: secsubtype before group (not after underlying)

Updated:
- src/decode.jl: Complete rewrite of InstrumentDefMsg decoder
- src/encode.jl: Complete rewrite of InstrumentDefMsg encoder
- src/messages.jl: Fix channel_id, leg_count, leg_index types to UInt16
- test/compatibility_utils.jl: Update JSON parsing for correct types

+++ src/messages.jl
@@ -422,7 +422,7 @@ struct InstrumentDefMsg
     appl_id::Int16
     maturity_year::UInt16
     decay_start_date::UInt16
-    channel_id::UInt8
+    channel_id::UInt16  # uint16_t per Databento docs
     currency::String
     settl_currency::String
     secsubtype::String
@@ -451,8 +451,8 @@ struct InstrumentDefMsg
     flow_schedule_type::Int8
     tick_rule::UInt8
     # New strategy leg fields in DBN v3
-    leg_count::UInt8
-    leg_index::UInt8
+    leg_count::UInt16  # uint16_t per Databento docs
+    leg_index::UInt16  # uint16_t per Databento docs
     leg_instrument_id::UInt32
     leg_raw_symbol::String
     leg_side::Side.T
